Improved type uniformity. Made sure not to mix doubles and floats, also modified npe_pdf_func to use std::round for pdfT0Sample to avoid accidentally truncating the value.

+++ include/DataStructures.h
@@ -6,7 +6,7 @@
 
 struct WaveformData {
 	unsigned short channel;
-	std::vector<float> waveform{};
+	std::vector<double> waveform{};
 };
 
 typedef struct {
@@ -38,8 +38,8 @@ typedef struct {
 
 typedef struct {
 	unsigned short ch;
-	float chi2ndf;
-	float baseline;
+	double chi2ndf;
+	double baseline;
 	std::vector<PEData> pes;
 } ChannelFitData;
